chore: remove old script and tests
