+++ index.html
@@ -7,7 +7,7 @@
     <title>ショート学習アプリ（完成版）</title>
     <link rel="stylesheet" href="styles.css">
     <!-- Firebase SDK -->
-    <script type="module" src="script.js"></script>
+    <script type="module" src="main.js"></script>
 </head>
 
 <body>
